+++ server/index.js
@@ -2,5 +2,7 @@
 'use strict';
 
 module.exports = function (app, info) {
+  delete require.cache[require.resolve('../lib/middleware')];
+
   require('../lib/middleware')(app, info)
 };